requirements.txt
absl-py==2.0.0
accelerate==0.25.0
adal==1.2.7
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==0.4.3
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
applicationinsights==0.11.10
argcomplete==2.1.2
asteval==0.9.31
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitarray==2.9.0
bitstring==4.1.4
blinker==1.7.0
cachetools==4.2.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
click-option-group==0.5.6
click-plugins==1.1.1
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
clip-anytorch==2.5.2
cloudpickle==2.2.1
cmake==3.27.9
colorama==0.4.6
comm==0.2.0
consistencydecoder @ git+https://github.com/openai/consistencydecoder.git@22a0449022f17a2d7bfc69535e8e8f3ff0585ecb
contextlib2==21.6.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
datasets==2.8.0
dctorch==0.1.2
debugpy==1.8.0
decorator==5.1.1
-e git+https://github.com/huggingface/diffusers.git@6bf1ca2c799f3f973251854ea3c379a26f216f36#egg=diffusers
dill==0.3.6
distro==1.8.0
docker==6.1.3
docker-pycreds==0.4.0
dotnetcore2==3.1.23
ecdsa==0.18.0
einops==0.3.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Executor==1.0.0
Flask-FontAwesome==0.1.5
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.1
ftfy==6.1.3
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-auth==2.25.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.62.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.59.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
humanize==4.9.0
idna==3.6
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflect==7.0.0
invisible-watermark==0.2.0
ipykernel==6.27.1
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
jsonmerge==1.9.2
jsonpickle==3.0.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
k-diffusion @ git+https://github.com/crowsonkb/k-diffusion.git@045515774882014cc14c1ba2668ab5bad9cbf7c0
keyring==24.3.0
kiwisolver==1.4.5
knack==0.10.1
kornia==0.6.0
lazy_loader==0.3
lit==17.0.6
lpips==0.1.4
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.14.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.0.0
multidict==6.0.4
multiprocess==0.70.14
ndg-httpsclient==0.5.1
nest-asyncio==1.5.8
networkx==3.2.1
ntlm-auth==1.5.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
openai==1.3.7
opencensus==0.11.3
opencensus-context==0.1.3
opencensus-ext-azure==1.1.12
opencensus-ext-logging==0.1.1
opencv-python==4.7.0.68
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==23.0
pandas==2.1.3
paramiko==3.4.0
parso==0.8.3
passpy==1.0.2
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.0.1
pkginfo==1.9.6
platformdirs==4.1.0
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-gnupg==0.5.2
python-multipart==0.0.6
pytorch-lightning==1.6.0
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.0
rich-click==1.7.2
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-image==0.20.0
scipy==1.11.4
seaborn==0.13.0
SecretStorage==3.3.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.38.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sshpubkeys==3.3.1
stack-data==0.6.3
starlette==0.27.0
streamlit==1.30.0
sympy==1.12
tabulate==0.9.0
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
taming-transformers-rom1504==0.0.6
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
test-tube==0.7.5
tifffile==2023.9.26
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torch-fidelity==0.3.0
torchaudio==2.0.2
torchdiffeq==0.2.3
torchmetrics==0.6.0
torchsde==0.2.6
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trampoline==0.1.2
transformers==4.36.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
urwid==2.3.4
uvicorn==0.24.0.post1
validators==0.22.0
wandb==0.16.1
watchdog==3.0.0
wcwidth==0.2.12
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
xdg==6.0.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
